reworked how shells and mods work
the loom pattern is now yielded a Shell::ShellApi object as its first
param, all capabilities are exposed through the ShellApi via loaded
Mods. Mods are responsible for providing every bit of capability to
the shell.

updated how mods are loaded and added aliasing and anonymous blocks to
mods. Loom::CoreMod::Exec was added to allow shell commands to be
executed.

* added mods Exec, Cd, Sudo
* addeed mod users
* simplified mod Files
* add hacsk to the shell for handling sudo better