broken state commit
up to this point i've been jumping through hoops trying to fix quoting
and escaping in commands for wrapping in arbitrary levels of other
commands and it's driving me insane. i'm throwing it all away and
trying a new approach (in future commits)

this commit is the current state of loom w/ a very manual and brittle
command escaping approach. it's only saved in case my new approach
totally flops